Interpolation type icons: tcb, linear, ease, const, clamped
All these icons are placed in the same sif file in different layers(groups)
and use time line/keyframes to tech to render final icon files. We are
animation software so why not take advance of time :) The most
important is that we can easily control icon to have same appearance
in color and shape when they are in same sif file.